[FIX][BE] Add CreatedBy field to Bucket and initialize it on creation (#53)
Add CreatedBy field to Bucket and initialize it on creation

The CreatedBy field has been added to the Bucket model to track the user's ID who created the bucket. This field is now set during bucket creation in the CreateBucket method. Additionally, its type was corrected to UUID and marked as non-null.